requirements.txt
addict==2.2.1
autopep8==1.4.4
captcha==0.3
celery==4.3.0
certifi==2019.6.16
coverage==4.5.3
exchangelib==1.12.5
Jinja2==2.10.1
nose2==0.9.1
Pillow==6.0.0
pylint==2.3.1
PyMySQL==0.9.3
qiniu==7.2.4
qrcode==6.1
redis==3.2.1
requests==2.22.0
requests-ntlm==1.1.0
shortuuid==0.5.0
tornado==6.0.2
xlwt==1.3.0
git+https://github.com/jie/dragonlib.git
wechatpy==1.8.4
